Add files via upload
This is the setup script for the database used in this project. 
In order to be executed successfully first create a database entitled BrassDuck. Then import the BrassDuckSampleData excel file into BrassDuck as [dbo].['fantasy-football-weekly-project$']. Then execute this script. It will create the tables, constraints, and populate the initial data held in the Players table.